Refactors file name
